refactor: Relocate close folder to management directory
