+++ app/(club)/club/management/close/content/headerSection.tsx
