+++ app/components/layout/header.tsx
@@ -45,7 +45,7 @@ const Header = () => {
     "/club/recruitment/create",
     "/notification",
     "/club/withdrawal",
-    "/club/close",
+    "/club/management/close",
     "/help",
     "/terms/user",
     "/terms/privacy",
